refactor(ratelimit): share host-refcount logic between guard and bypass
BackgroundRateLimitGuard and InteractiveRateLimitBypass were byte-for-byte
identical ConcurrentHashMap<String, Int> refcounters with different names.
Extract the shared HostRefCounter so a future fix to the refcount/race
logic only needs to happen once. Public API and behavior unchanged.
